Changed implementation for rules implementation to inherit from Rule class
